requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.1.0
audioread==3.0.0
babel==2.12.1
beautifulsoup4==4.12.2
black==19.10b0
braceexpand==0.1.7
cachecontrol==0.13.1
cachetools==5.3.1
cachy==0.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.0.2
click-default-group==1.2.4
clikit==0.6.2
cmake==3.27.4.1
colorama==0.4.6
contourpy==1.1.0
crashtest==0.3.1
cryptography==41.0.3
cycler==0.11.0
datasets==2.14.5
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.20.1
editdistance==0.6.2
einops==0.6.1
ensureconda==1.4.3
exceptiongroup==1.1.3
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gdown==4.7.1
gitdb==4.0.10
gitpython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
html5lib==1.1
huggingface-hub==0.17.1
hydra-core==1.3.2
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
inflect==7.0.0
iniconfig==2.0.0
isort==5.12.0
jaraco-classes==3.3.0
jeepney==0.8.0
jinja2==3.1.2
jiwer==2.5.2
joblib==1.3.2
keyring==24.2.0
kiwisolver==1.4.5
kornia==0.7.0
latexcodec==2.0.1
lazy-loader==0.3
levenshtein==0.21.1
librosa==0.10.1
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.40.1
loguru==0.7.2
markdown==3.4.4
markdown-it-py==3.0.0
markdown2==2.4.10
markupsafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.3
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numba==0.57.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
packaging==23.1
pandas==2.1.0
parameterized==0.9.0
pastel==0.2.1
pathspec==0.11.2
pathtools==0.1.2
pillow==10.0.0
pkginfo==1.9.6
plac==1.3.5
platformdirs==3.10.0
pluggy==1.3.0
pooch==1.7.0
protobuf==4.24.3
psutil==5.9.5
pyannote-core==5.0.0
pyannote-database==5.0.1
pyannote-metrics==3.2.1
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
pygments==2.16.1
pylev==1.4.0
pyparsing==3.1.1
pysocks==1.7.1
pytest==7.4.2
pytest-runner==6.0.0
python-dateutil==2.8.2
pytorch-lightning==2.0.7
pytz==2023.3.post1
pyyaml==6.0.1
rapidfuzz==2.13.7
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
ruamel-yaml==0.17.32
ruamel-yaml-clib==0.2.7
sacremoses==0.0.53
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
secretstorage==3.3.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.30.0
setproctitle==1.3.2
setuptools==65.5.1
shellingham==1.5.3
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
soxr==0.3.6
sphinx==7.2.5
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-bibtex==2.6.1
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
text-unidecode==1.3
texterrors==0.4.4
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
torchmetrics==1.1.2
tqdm==4.66.1
transformers==4.36.0
typed-ast==1.5.5
typer==0.9.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.5
wandb==0.15.10
webdataset==0.1.62
webencodings==0.5.1
werkzeug==2.3.7
wget==3.2
wrapt==1.15.0
xxhash==3.3.0
yarl==1.9.2
youtokentome==1.0.6
zipp==3.16.2